requirements.txt
Django~=4.2.1
mysql~=0.0.3
mysql-connector-python~=8.0.33
django-environ~=0.10.0